+++ modules/Preferences.py
@@ -1,4 +1,4 @@
-import Utils
+from modules import Utils
 import logging
